Added a logic to attach the existing code around the sandbox.
This way the application can directly talk to webworks because iframes do nto get access to the "Blackberry" object.

Also renamed history because of conflicts in various places, and added the ability to quit the app in the blackberry.

Made all of the links invoke the browser on the playbook rather than nav inside the application

Made it so that the application will connect to the correct remote.js on @rem's server for now so that you can actually use this as a host for another JS program.